requirements.txt
bleach==6.2.0
bokeh==3.7.2
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.57.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.8
linkify-it-py==2.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdit-py-plugins==0.4.2
mdurl==0.1.2
narwhals==1.34.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
panel==1.6.2
param==2.2.0
pillow==11.1.0
plotly==6.0.1
pluggy==1.5.0
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyserial==3.5
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyviz_comms==3.0.4
PyYAML==6.0.2
requests==2.32.3
setuptools==78.1.0
shiboken2==5.13.2
six==1.17.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.3.0
webencodings==0.5.1
xyzservices==2025.1.0
